+++ db/migrate/20211119120347_add_title_to_content_templates.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddTitleToContentTemplates < ActiveRecord::Migration[6.1]
+  def change
+    add_column :folio_content_templates, :title, :string
+  end
+end